move manager in folder
