+++ selfdrive/loggerd/tests/test_loggerd.py
@@ -14,7 +14,7 @@ from cereal.services import service_list
 from common.basedir import BASEDIR
 from common.timeout import Timeout
 from common.params import Params
-import selfdrive.manager as manager
+import selfdrive.manager.manager as manager
 from selfdrive.hardware import TICI, PC
 from selfdrive.loggerd.config import ROOT
 from selfdrive.test.helpers import with_processes
