+++ selfdrive/test/helpers.py
@@ -6,7 +6,7 @@ from nose.tools import nottest
 from selfdrive.hardware.eon.apk import update_apks, start_offroad, pm_apply_packages, android_packages
 from selfdrive.hardware import PC
 from selfdrive.version import training_version, terms_version
-from selfdrive.manager import start_managed_process, kill_managed_process, get_running
+from selfdrive.manager.manager import start_managed_process, kill_managed_process, get_running
 
 def set_params_enabled():
   from common.params import Params
