+++ tools/sim/launch_openpilot.sh
@@ -5,4 +5,4 @@ export NOBOARD="1"
 export SIMULATION="1"
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
-cd ../../selfdrive && ./manager.py
+cd ../../selfdrive/manager && ./manager.py
